+++ index.js
@@ -1,6 +1,3 @@
-export {default as color} from "./src/color";
-export {default as rgb} from "./src/rgb";
-export {default as hsl} from "./src/hsl";
-export {default as lab} from "./src/lab";
-export {default as hcl} from "./src/hcl";
+export {default as color, rgb, hsl} from "./src/color";
+export {default as lab, hcl} from "./src/lab";
 export {default as cubehelix} from "./src/cubehelix";
