+++ src/math.js
@@ -1,0 +1,2 @@
+export var deg2rad = Math.PI / 180;
+export var rad2deg = 180 / Math.PI;
